bot/requirements.txt
click
pyyaml
python-dotenv
aiogram
grpcio
grpcio-tools
